+++ libsemanage/src/parse_utils.c
@@ -85,7 +85,7 @@ int parse_skip_space(semanage_handle_t * handle, parse_info_t * info)
 
 	size_t buf_len = 0;
 	ssize_t len;
-	int lineno = info->lineno;
+	unsigned int lineno = info->lineno;
 	char *buffer = NULL;
 	char *ptr;
 
@@ -271,7 +271,7 @@ int parse_fetch_string(semanage_handle_t * handle,
 {
 
 	const char *start = info->ptr;
-	int len = 0;
+	size_t len = 0;
 	char *tmp_str = NULL;
 
 	if (parse_assert_noeof(handle, info) < 0)
